requirements.txt
# Dependencias principales
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.23.0

# Datos financieros
yfinance>=0.2.12

# Web scraping y búsqueda
requests>=2.28.0
beautifulsoup4>=4.11.0
duckduckgo-search>=3.0.0
tavily-python

# Visualización de datos
plotly>=5.13.0
matplotlib>=3.6.0

# Utilidades adicionales
python-dateutil>=2.8.0
pytz>=2022.1
tqdm>=4.64.0
openai
ta
scikit-learn
scipy
python-dotenv
logging
tenacity
seaborn
statsmodels
markdown>=3.4.0
pdfkit>=1.0.0

# Base de datos y correo
mysql-connector-python>=8.0.0
watchdog>=2.1.0

# Nota: smtplib, email e importlib son parte de la biblioteca estándar de Python